bot/requirements.txt
hikari==2.0.0.dev121
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorlog==6.7.0
frozenlist==1.4.0
h11==0.14.0
hikari-lightbulb==2.3.3
hikari-miru==3.1.2
httpcore==0.17.3
httpx==0.24.1
idna==3.4
multidict==6.0.4
pydantic>=1.8.0,<2.0.0
pocketbase==0.9.1
sigparse==3.0.0
sniffio==1.3.0
typing_extensions==4.7.1
yarl==1.9.2